Add sponsors
Doing away with the gridism framework as we won't need much of it.

Instructions to add a new sponsor:

To add a new sponsor, just add an anchor tag similar to one of the
existing ones inside a list tag. apply a class of .imagereplace for
the anchor tags. Add an id to the anchor tag and in the css, give the
id a background of the company logo and specify the width. No need to
specify a height. All the images should be resized to fit a height of
60px.