Hardcode the names and links back in the posts
Jekyll's scope is a little weird. The page object is the global template
object that resides in the root (in this case, the blog/index.html where
    the list of the blogposts are shown) whereas the page object refers
to the YAML frontmatter of the template getting embedded inside the root
page. Better to just hardcode the names since we are writing the names
in the title of the post anyway.